Cleanup dead code
