+++ src/main/java/com/github/cfogrady/dim/modifier/data/AppState.java
@@ -4,7 +4,6 @@ package com.github.cfogrady.dim.modifier.data;
 import com.github.cfogrady.dim.modifier.data.card.CardData;
 import com.github.cfogrady.dim.modifier.data.card.Character;
 import com.github.cfogrady.dim.modifier.data.firmware.FirmwareData;
-import com.github.cfogrady.dim.modifier.view.BemInfoView;
 import com.github.cfogrady.vb.dim.card.Card;
 import com.github.cfogrady.vb.dim.sprite.SpriteData;
 import lombok.Data;
@@ -21,7 +20,6 @@ public class AppState {
     private Card rawCard;
     private CardData<?, ?> cardData;
     private boolean safetyModeOn;
-    private BemInfoView currentView;
     private File lastOpenedFilePath;
     private int selectedBackgroundIndex;
