--- src/main/java/com/github/cfogrady/dim/modifier/view/BemSpritesView.java
@@ -1,4 +1,0 @@
-package com.github.cfogrady.dim.modifier.view;
-
-public class BemSpritesView {
-}
