+++ src/main/java/com/github/cfogrady/dim/modifier/view/controller/CharacterViewController.java
@@ -5,7 +5,6 @@ import com.github.cfogrady.dim.modifier.SpriteReplacer;
 import com.github.cfogrady.dim.modifier.controls.ImageIntComboBox;
 import com.github.cfogrady.dim.modifier.data.AppState;
 import com.github.cfogrady.dim.modifier.data.card.Character;
-import com.github.cfogrady.dim.modifier.view.BemCharactersView;
 import com.github.cfogrady.vb.dim.sprite.SpriteData;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
@@ -16,12 +15,14 @@ import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.StackPane;
+import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 import java.net.URL;
 import java.util.ResourceBundle;
 import java.util.Timer;
+import java.util.TimerTask;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -52,7 +53,7 @@ public class CharacterViewController implements Initializable {
     }
 
     private int characterSelection = 0;
-    private BemCharactersView.NameUpdater nameUpdater;
+    private NameUpdater nameUpdater;
     private SubViewSelection subViewSelection = SubViewSelection.STATS;
 
 
@@ -88,7 +89,7 @@ public class CharacterViewController implements Initializable {
         ImageView imageView = new ImageView(image);
         imageView.setViewport(new Rectangle2D(0, 0, 80, 15));
         if(nameSprite.getWidth() > 80) {
-            nameUpdater = new BemCharactersView.NameUpdater(nameSprite.getWidth(), imageView, -80);
+            nameUpdater = new NameUpdater(nameSprite.getWidth(), imageView, -80);
             timer.scheduleAtFixedRate(nameUpdater, 0, 33);
         }
         //nameBox.setBackground();
@@ -137,6 +138,26 @@ public class CharacterViewController implements Initializable {
                 transformationViewController.refreshAll();
                 subView.getChildren().add(transformationsSubView);
             }
+        }
+    }
+
+    @AllArgsConstructor
+    public static class NameUpdater extends TimerTask {
+        private final int spriteWidth;
+        private final ImageView imageView;
+        private int offsetX;
+        @Override
+        public void run() {
+            // hack to make it pause at the start and end.
+            if(offsetX > spriteWidth) {
+                offsetX = -80;
+                imageView.setViewport(new Rectangle2D(0, 0, 80, 15));
+            } else if(offsetX < 0) {
+                imageView.setViewport(new Rectangle2D(0, 0, 80, 15));
+            } else {
+                imageView.setViewport(new Rectangle2D(offsetX % spriteWidth, 0, 80, 15));
+            }
+            offsetX += 1;
         }
     }
 }
